ecommerce-backend/requirements.txt
# Core
Django>=5.0
djangorestframework>=3.15
psycopg2-binary>=2.9
python-dotenv>=1.0
dj-database-url>=2.1
bleach>=6.1.0  # For input sanitization

# Auth & filtering
djangorestframework-simplejwt
drf-yasg
django-filter
django-cors-headers>=4.3

# Images
pillow

# Production
gunicorn>=21.2
whitenoise>=6.7

# Add these to requirements.txt
pytest>=7.4.0
pytest-django>=4.5.2
pytest-cov>=4.1.0